+++ application/models/EmailHandler.php
@@ -3271,6 +3271,7 @@ class EmailHandler extends FatModel
         $langId = $this->commonLangId;
         $vars = [
             '{sender_name}' => $giftcard['user_name'],
+            '{sender_email}' => $giftcard['credential_email'],
             '{recipient_name}' => $giftcard['ogcards_receiver_name'],
             '{recipient_email}' => $giftcard['ogcards_receiver_email'],
             '{giftcard_code}' => $giftcard['ogcards_code'],
@@ -3298,9 +3299,9 @@ class EmailHandler extends FatModel
         $langId = $this->commonLangId;
         $vars = [
             '{sender_name}' => $giftcard['user_name'],
+            '{sender_email}' => $giftcard['credential_email'],
             '{recipient_name}' => $giftcard['ogcards_receiver_name'],
             '{giftcard_code}' => $giftcard['ogcards_code'],
-            '{contact_us_email}' => FatApp::getConfig('CONF_CONTACT_EMAIL')
         ];
         $sendEmail = false;
         if (!empty($giftcard['ogcards_receiver_email'])) {