+++ CHANGELOG.md
@@ -81,6 +81,7 @@
     - Bug-085674 - Distance-based Delivery: Same category listing appears multiple times in the categories collection on the homepage
     - Bug-087073 - The content issue in the email change request email.
     - Bug-087200 - partial payment is not working with wallet + stripe
+    - Bug-086479 - sender email should also be visible to receiver and admin.
 
 ## Hit URL:
     - admin/patch-update/updateProductRating